Rethink how hints are stored
The Game now has two Nonograms in it, the "picture" which is the
immutable thing we initialized the puzzle with and the "puzzle" which is
the initially blank slate the player is trying to get to look like the
picture.

Each Nonogram stores hints, calculated at load time.

Future work is to update the hints for the puzzle when the player makes
an edit so we can compare them against the hints for the picture.